박자,마디 계산 오류 수정

+++ Assets/Scripts/Map Editor/Note/MapEditorMeterChangerNote.cs
@@ -44,7 +44,7 @@ public class MapEditorMeterChangerNote : MapEditorNote
         {
             whenSummonBeat = beat,
             beatPerBar = _meter1,
-            beatLengthRate = _meter1 / _meter2
+            beatLengthRate = (float)_meter1 / _meter2
         };
     }
 }
